onlineshop/requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.11
Pillow==2.8.2
static3==0.7.0
psycopg2==2.6.1
gunicorn==19.6.0
